separate model dicrectory according to domain
