+++ lib/models/common/human_life.dart
@@ -1,5 +1,5 @@
-import 'package:HumanLifeGame/models/life_road.dart';
-import 'package:HumanLifeGame/models/user.dart';
+import 'package:HumanLifeGame/models/common/life_road.dart';
+import 'package:HumanLifeGame/models/common/user.dart';
 
 class HumanLifeModel {
   HumanLifeModel({this.title, this.author, this.lifeRoad});
