+++ lib/models/common/life_event_doc.md
