+++ lib/models/common/life_item.dart
