+++ lib/models/common/life_road.dart
@@ -1,4 +1,4 @@
-import 'package:HumanLifeGame/models/life_step.dart';
+import 'package:HumanLifeGame/models/common/life_step.dart';
 import 'package:flutter/foundation.dart';
 
 class LifeRoadModel {
