+++ lib/models/common/life_step.dart
@@ -1,4 +1,4 @@
-import 'package:HumanLifeGame/models/life_event.dart';
+import 'package:HumanLifeGame/models/common/life_event.dart';
 import 'package:flutter/foundation.dart';
 
 class LifeStepModel {
