+++ lib/screens/play_room/dice_result.dart
@@ -1,4 +1,4 @@
-import 'package:HumanLifeGame/models/player_action.dart';
+import 'package:HumanLifeGame/models/play_room/player_action.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
