+++ lib/screens/play_room/player_action.dart
@@ -1,5 +1,5 @@
 import 'package:HumanLifeGame/i18n/i18n.dart';
-import 'package:HumanLifeGame/models/player_action.dart';
+import 'package:HumanLifeGame/models/play_room/player_action.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
